% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{Entropy}
\alias{Entropy}
\title{Function that calculates the Entropy of a chemical reaction}
\usage{
Entropy(dG, dH, Temperature)
}
\arguments{
\item{dG}{Gibbs free energy of a chemical reaction}

\item{dH}{Enthalpy of a chemical reaction}

\item{Temperature}{Temperature in celcius}
}
\value{
Entropy of a chemical reaction
}
\description{
Function that calculates the Entropy of a chemical reaction
}
