% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{Gibbs.free.energy}
\alias{Gibbs.free.energy}
\title{Function that calculates the Gibbs free energy of a chemical reaction}
\usage{
Gibbs.free.energy(K, Temperature, R = 0.0019872)
}
\arguments{
\item{K}{Molar affinity constant of a ligand binding to a macromolecule}

\item{Temperature}{Temperature in celcius}

\item{R}{Gas constant. Default = 0.0019872 kcal/mol/K}
}
\value{
Gibbs free energy of a chemical reaction
}
\description{
Function that calculates the Gibbs free energy of a chemical reaction
}
