% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{Kd.app.calc}
\alias{Kd.app.calc}
\title{Function that calculates the mM apparent disassociation constant (Kd)}
\usage{
Kd.app.calc(
  metabolite = "Glutathione",
  pH = 7.5,
  I = 0.165,
  M.charge = 2,
  A = 0.524,
 
    constants.path = "Binding_constant_concentration_data/210525_Metaboites_binding_Mg_thermodynamics.csv"
)
}
\arguments{
\item{metabolite}{The metabolite you want to model an apparant Kd for}

\item{pH}{The pH you want to model an apparant Kd for}

\item{I}{The ionic strength you want to model a Kd for}

\item{M.charge}{The charge of the metal ion}

\item{A}{Constant for SIT at 25 degC}

\item{constants.path}{Path to the file containing critical stability constants for metal ion and proton binding}
}
\value{
An apparant Kd in mM (M/1000)
}
\description{
Corrects for ionic strength using the using Specific Interaction Theory (SIT)(1) and corrects for
pH using pKa's to calculate populations of metal ion binding competent protonation states(2).
}
\details{
\link{1} 	Scatchard, George. Concentrated Solutions of Strong Electrolytes. Chem. Rev. 1936, 19 (3), 309–327. https://doi.org/10.1021/cr60064a008.
\link{2} 	Mattocks, J. A.; Tirsch, J. L.; Cotruvo, J. A. Chapter Two - Determination of Affinities of Lanthanide-Binding Proteins Using Chelator-Buffered Titrations. In Methods in Enzymology; Cotruvo, J. A., Ed.; Rare-Earth Element Biochemistry: Characterization and Applications of Lanthanide-Binding Biomolecules; Academic Press, 2021; Vol. 651, pp 23–61. https://doi.org/10.1016/bs.mie.2021.01.044.
}
