% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{log10K.0.calculator}
\alias{log10K.0.calculator}
\title{Function that calculates the log10 of a metal ion affinity constant for Ionic strength = 0 using the Debye-Huckel equation}
\usage{
log10K.0.calculator(log10K, I, M.charge, X.charge, XM.charge, A = 0.524)
}
\arguments{
\item{log10K}{Molar affinity constant of a metabolite binding to a metal ion}

\item{I}{ionic strength}

\item{M.charge}{charge of the metal ion}

\item{X.charge}{charge of the metabolite}

\item{A}{Debye-Huckel constant for the given temperature. Default A = 0.524 for 25C.}

\item{MX.charge}{Charge of the metal ion-metabolite complex}
}
\value{
the log10K at an ionic strength of zero
}
\description{
Function that calculates the log10 of a metal ion affinity constant for Ionic strength = 0 using the Debye-Huckel equation
}
