% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{log10K.IS.calculator.2}
\alias{log10K.IS.calculator.2}
\title{Function that calculates the log10 of a metal ion affinity constant for any Ionic strength  using the Debye-Huckel equation}
\usage{
log10K.IS.calculator.2(
  log10K.ref,
  I.ref,
  I,
  M.charge,
  X.charge,
  XM.charge,
  A = 0.51
)
}
\arguments{
\item{log10K.ref}{Molar affinity constant of a metabolite binding to a metal ion that you are referencing}

\item{I.ref}{Ionic strength of the reference constant}

\item{I}{ionic strength for the condition you want to calculate}

\item{M.charge}{charge of the metal ion}

\item{X.charge}{charge of the metabolite}

\item{XM.charge}{Charge of the metal ion-metabolite complex}

\item{A}{Debye-Huckel constant for the given temperature. Default A = 0.524 for 25C.}
}
\value{
the log10K at an ionic strength of zero
}
\description{
Function that calculates the log10 of a metal ion affinity constant for any Ionic strength  using the Debye-Huckel equation
}
