% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboMgITC.R
\name{MetaboMgITC}
\alias{MetaboMgITC}
\title{Generalized ITC data fitting function}
\usage{
MetaboMgITC(
  cell,
  blank,
  Thermodynamic.equation = "Wiseman.isotherm",
  Fit.start = list(H = 2.534, K = 2880),
  Remove.injection = 1,
  Saturation.threshold = FALSE,
  Save.path.and.prefix = FALSE
)
}
\arguments{
\item{cell}{Object created by "read.itc" for the ITC run containing the macromolecule.}

\item{blank}{Object created by "read.itc" for the ITC run containing the ligand tirated into buffer.}

\item{Thermodynamic.equation}{The thermodynamic model that you want to fit the data to. Default = "wiseman.isotherm".}

\item{Fit.start}{Starting parameters for the non-linear regression. Default = = list(H = 2.534, K = 2880).}

\item{Remove.injection}{Injections you want to remove from subsequent analysis. Default = 1 is the standard for ITC experiments. More than one injection can be supplied in a vector.}

\item{Saturation.threshold}{For low c value ITC experiments fit to Wiseman isotherms, macomolecule saturation ranging from 70\% to 90\% produce the same answers. Thus, sometimes it is desirable to standardize the saturation threshold between experiments to minimize degredation of a ligand or macromolecule (for example with ATP binding Mg). Default = FALSE. If set to 0.8, this function will only fit data required to reach 80\% macromolecule saturation.}
}
\value{
A list containing summary statistics for the fit and a graphical depiction of the fits, the raw ITC curve, and the summary statistics.
}
\description{
Fits an ITC experiment to a Wiseman isotherm(1) to determine thermodynamic statistics. Raw ".itc" formatted data
should be read in using "read.itc". Fits the data to a thermodynamic model then prints a graphical summary of the fit.
}
\details{
(1) Turnbull, W. B.; Daranas, A. H. On the Value of c: Can Low Affinity Systems Be Studied by Isothermal Titration Calorimetry? Journal of the American Chemical Society 2003, 125 (48), 14859–14866. https://doi.org/10.1021/ja036166s.
}
