% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{pKa.IS.calculator}
\alias{pKa.IS.calculator}
\title{Function that calculates the pKa of any Ionic strength using the Debye-Huckel equation}
\usage{
pKa.IS.calculator(pKa.ref, I.ref, I, X.charge, HX.charge, A = 0.524)
}
\arguments{
\item{pKa.ref}{pKa for the reference that you are referencing}

\item{I.ref}{Ionic strength of the reference constant}

\item{I}{ionic strength for the condition you want to calculate}

\item{X.charge}{charge of the base}

\item{HX.charge}{Charge of acid}

\item{A}{Debye-Huckel constant for the given temperature. Default A = 0.524 for 25C.}
}
\value{
the log10K at an ionic strength of zero
}
\description{
Function that calculates the pKa of any Ionic strength using the Debye-Huckel equation
}
