% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting_functions.R
\name{read.itc}
\alias{read.itc}
\title{Imports a VP-ITC instument ".itc" formatted data file into R}
\usage{
read.itc(
  file.concentrations,
  print.peak.integration.graph = FALSE,
  integration.folder = NA
)
}
\arguments{
\item{file.concentrations}{Vector containing paths to an ITC file then reagent concentration ordered c(Path, syringe X concentration, syringe M concentration, syringe C concentration, cell X concentration, cell M concentration, cell C concentration). Enter a file path to use the concentration embedded in the raw .itc file.}

\item{print.peak.integration.graph}{Print an aesthetic depiction of the peak integration for each injection. Options = TRUE or FALSE. Default = FALSE. If TRUE will print a png in your working directory.}
}
\value{
A list of two data frames that can be passed to data analysis functions like "MetaboMgITC". The first dataframe contains the raw differential power (microcalories per second) as a function of time (second). The second data frame contains the heat of each injection (kilocalorie per mole injectant) and reagent concentration (molar). Volumes are reported as Liters. M is the concentration of the reagent in the cell. X and dX are the concentration of the ligand in the cell and the moles of ligand added in each injection respectively. dQ and dQ.dX are the heat produced by each injection and the heat produced by each injection divided by the moles added by each injection.
}
\description{
Opens a MicroCal VP-ITC instrument ".itc" formatted data file into R. Parses raw text into convenient data
frames. Calculates reagent concentrations. Integrates raw differential power curves to determine the heat
of each injection using the polyarea function in pracma.
}
