% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{saturation}
\alias{saturation}
\title{Function that calculates the saturation of a macromolecule in an ITC experiment}
\usage{
saturation(K, M, X, V, V0)
}
\arguments{
\item{K}{Molar affinity constant of a ligand binding to a macromolecule}

\item{M}{Macromolecule in a cell}

\item{X}{Ligand concentration in a cell}

\item{V}{Cell volume in Liters}

\item{V0}{Original cell volume}
}
\value{
Percent saturation of the macromolecule by bound ligand
}
\description{
Function that describes heat produced by each injection in an ITC experiment when there is 1 to 1 binding of ligand to a macromolecule.
}
