% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermodynamic_equations.R
\name{Wiseman.isotherm.competative}
\alias{Wiseman.isotherm.competative}
\title{Function that describes heat produced by each injection in an ITC experiment}
\usage{
Wiseman.isotherm.competative(
  V,
  X,
  M,
  C,
  H = 2534,
  K = 2880,
  Hc = 2534,
  Kc = 2880,
  c = 0,
  h = 0
)
}
\arguments{
\item{V}{Cell volume in Liters}

\item{X}{Ligand concentration in a cell}

\item{M}{Macromolecule in a cell}

\item{H}{Enthalpy of a ligand binding to a macromolecule}

\item{K}{Molar affinity constant of a ligand binding to a macromolecule}
}
\value{
Heat produced by an injection
}
\description{
Function that describes heat produced by each injection in an ITC experiment when there is 1 to 1 binding of ligand to a macromolecule and a competator.
}
